+++ src/main/java/com/alzzaipo/member/application/service/account/LocalAccountService.java
@@ -5,9 +5,9 @@ import com.alzzaipo.common.LoginType;
 import com.alzzaipo.common.email.domain.Email;
 import com.alzzaipo.common.email.domain.EmailVerificationCode;
 import com.alzzaipo.common.email.domain.EmailVerificationPurpose;
-import com.alzzaipo.common.email.port.out.smtp.SendEmailVerificationCodePort;
-import com.alzzaipo.common.email.port.out.verification.CheckEmailVerificationCodePort;
-import com.alzzaipo.common.email.port.out.verification.SaveEmailVerificationCodePort;
+import com.alzzaipo.common.email.application.port.out.smtp.SendEmailVerificationCodePort;
+import com.alzzaipo.common.email.application.port.out.verification.CheckEmailVerificationCodePort;
+import com.alzzaipo.common.email.application.port.out.verification.SaveEmailVerificationCodePort;
 import com.alzzaipo.common.exception.CustomException;
 import com.alzzaipo.common.token.TokenUtil;
 import com.alzzaipo.common.token.application.port.out.SaveRefreshTokenPort;
