+++ src/main/java/com/alzzaipo/notification/application/service/NotificationService.java
@@ -4,10 +4,10 @@ import com.alzzaipo.common.Id;
 import com.alzzaipo.common.email.domain.Email;
 import com.alzzaipo.common.email.domain.EmailVerificationCode;
 import com.alzzaipo.common.email.domain.EmailVerificationPurpose;
-import com.alzzaipo.common.email.port.out.smtp.SendEmailVerificationCodePort;
-import com.alzzaipo.common.email.port.out.verification.CheckEmailVerificationCodePort;
-import com.alzzaipo.common.email.port.out.verification.DeleteEmailVerificationStatusPort;
-import com.alzzaipo.common.email.port.out.verification.SaveEmailVerificationCodePort;
+import com.alzzaipo.common.email.application.port.out.smtp.SendEmailVerificationCodePort;
+import com.alzzaipo.common.email.application.port.out.verification.CheckEmailVerificationCodePort;
+import com.alzzaipo.common.email.application.port.out.verification.DeleteEmailVerificationStatusPort;
+import com.alzzaipo.common.email.application.port.out.verification.SaveEmailVerificationCodePort;
 import com.alzzaipo.common.exception.CustomException;
 import com.alzzaipo.member.application.port.in.member.FindMemberAccountEmailsQuery;
 import com.alzzaipo.notification.application.port.dto.ChangeNotificationEmailCommand;
